Create Map Barrel Component
Create a component that loads the Map component only when window is
defined. This prevents any part of the Map component from being loaded
or having any of its code executed by the server rather than the
client. This is needed to prevent any issues that leaflet would have if
run outside of the client side.