Add loading animation to Map
Add a loading animation to the Map component. This will show a loading
spinner as projects or deployments are being added to the map as
markers. This prevents an empty map from being displayed for an extended
period of time as the user is waiting for markers to be loaded.