Fix no selected species behavior
Fix the case where no species are selected when submitting search
filters for the deployment search. This properly checks if the length of
the array of selected species is greater than 0 before applying the
filter. This avoids a situation where the query is filtering on an empty
array, which would return no deployments.