Handle no deployments in getDeploymentsCsv query
Handle the case where an empty array is passed in place of a list of
deployment nids. This prevents the express app from crashing and
throwing a SQL syntax error when binding the empty array to the query as
a parameter value. Respond with empty array if no deployments are
listed.