Show download link as Button
Show the CSV download link as a button. While the deployment data is
loading, show a loading spinner on the button. This will make it much
more clear to the user that the link is interactable. The loading state
indicates that the button is not ready to be clicked yet.