Pass value and label among filter components
Rather than storing the project ids and species names to show filter
options, store the React Multiselect form of them, including label and
value. This allows more information to be passed to
DeploymentMultiselect to set default values to be selected. Use
structure of { value, label } in the url search parameters. Modify the
express getAllDeployments query to parse these objects to get their
values.