Default deployments page
Set the default behavior for the deployments page. If there are no
filter parameters set, then do not load or display any deployments. This
allows the deployments page, as well as the index, to be used as the
entry point by which users would apply filters and search for
deployments.