Programatically add rectangle as layer on init
When initializing the map and loading the page, programatically add the
rectangle as a layer on the map rather than using the react leaflet draw
rectangle component. This allows the initial rectangle to follow the
same logic as following drawn rectangles. The layer can be removed upon
a new layer's creation. Remove the rectBounds prop from GeoFilterMap
becuse it is no longer used to check if we need to conditionally render
the Rectangle component.