Revert to layer ref
Instead of switching to a state variable, use a ref to reference the
rectLayer. This prevents issues where it is likely that the
EditFilterLayer component was unmounting as the layer was being
reassigned a value. Instead, pass a function that explicitly reassigns
rectLayer.current so that the value can be updated though passed through
props. This prevents an issue where drawing a new rectangle does not
delete the old one every time.