Do not use chart title in data
Instead of setting a title in the data, set the species. If using the
title keyword, the chart will show the title when hovering. To prevent
this, do not give each segment a title. Call the species name species so
that the species name can be accessed when setting the tooltip content.