Set multiselect to allow scrolling overflow
Set the maxHeights of the child elements of the multiselect to 36px to
match the container size and prevent unnecessary scroll bars. When
selecting long-named or many values, the user can scroll to read all the
selected options. This also prevents the selected values from causing
the multiselect not to shrink when resizing.