Use only searched filters
When downloading the sequence data, only use filters that have been
applied through the search button. Do not use filters that have been
set, but have not yet been searched. This prevents confusing behavior
where a set of deployments may be showing, but messages stating that
there is no sequence data when the user has applied different filters to
the species, projects, or geographical filter fields.